requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.1.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==26.1.0
beautifulsoup4==4.14.3
bitarray==3.8.1
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
ckzg==2.1.7
click==8.3.3
contourpy==1.3.3
curl_cffi==0.15.0
cycler==0.12.1
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
flatbuffers==25.12.19
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.47
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.16.0
hexbytes==1.3.1
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.14.0
kiwisolver==1.5.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.7.1
multitasking==0.0.12
namex==0.1.0
narwhals==2.20.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.1
pandas==3.0.2
parsimonious==0.10.0
patsy==1.0.2
peewee==4.0.4
pillow==12.2.0
platformdirs==4.9.6
propcache==0.4.1
protobuf==4.25.9
pyarrow==24.0.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyunormalize==17.0.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rlp==4.1.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
statsmodels==0.14.6
streamlit==1.56.0
tenacity==9.1.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.5
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
web3==7.15.0
websockets==15.0.1
Werkzeug==3.1.8
wrapt==2.1.2
yarl==1.23.0
yfinance==1.3.0
